Change WebSocketServer to bind correctly on IP.
Change WebSocketServer constructor to take IPAddress object and int port rather than a location string.
Adjust WebSocketServer members as we now store the ip/port in an IPEndPoint.

WebSocketServer constructor no longer accepts a "location" string which was being parsed to know if the server was "ws" or "wss" (secure). Instead we now look to whether a secure Certificate has been set (which was being checked in addition to the "wss" anyway).